pyproject.toml
[project]
name = "luppita"
version = "0.1.0"
description = "Luppita — AI property management agent for Colombian landlords"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.29.0",
    "a2ui-agent-sdk>=0.2.1",
    "python-dotenv>=1.2.2",
    "google-api-python-client>=2.0.0",
    "google-auth>=2.0.0",
    "google-cloud-logging>=3.0.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "ruff>=0.15.10",
]
